Solved compiler warnings under GCC 3.4.x: errno is now slong. Fixed various signedness warnings and unhandled enum values.
git-svn-id: https://freenos.googlecode.com/svn/trunk@79 824000e8-63cf-11de-a6e1-e33d94ba4d19

+++ lib/libposix/unistd.cpp
@@ -45,7 +45,7 @@ ssize_t read(int fildes, void *buf, size_t nbyte)
     errno = fs.result;
     
     /* Success. */
-    return errno == ESUCCESS ? fs.size : -1;
+    return errno == ESUCCESS ? fs.size : (ssize_t) -1;
 }
 
 ssize_t write(int fildes, const void *buf, size_t nbyte)
@@ -60,7 +60,7 @@ ssize_t write(int fildes, const void *buf, size_t nbyte)
     errno = fs.result;
     
     /* Give the result back. */
-    return errno == ESUCCESS ? fs.size : -1;
+    return errno == ESUCCESS ? fs.size : (ssize_t) -1;
 }
 
 int close(int fildes)
